refactor(agent): split openclaw/manager.rs into focused submodules
`openclaw/manager.rs` was a 2,830-line file mixing the OpenClawManager
struct/impl, the tungstenite WebSocket loop, identity loading, Ed25519
signing, URL/IP validation, exponential backoff, and ~1,090 lines of
integration tests. Split it into focused siblings, all production files
now under 500 lines:

- manager.rs (315) - OpenClawManager struct + public API surface
- connection.rs (454) - WebSocket session/reconnect loop
- runtime_state.rs (187) - per-gateway runtime snapshot mutators
- identity.rs (241) - device identity loading + filesystem fallback
- device_proof.rs (202) - Ed25519 signing + payload + scope validation
- url_validation.rs (150) - URL parsing, DNS resolution, IP allowlist
- dto.rs (110) - public DTOs (status, snapshot, view, requests, events)
- command.rs (82) - `openclaw` external binary invocation
- backoff.rs (40) - reconnect attempt + jittered sleep helpers
- util.rs (34) - shared scalar helpers (now_ms, normalize_*)
- session.rs (26) - internal channel types
- tests.rs (1116) - integration tests (tests-only, may exceed 600)

`mod.rs` declares the new submodules and preserves the existing public
re-exports. `manager::GatewayView` is re-exported on the manager path so
`api_server.rs` `crate::openclaw::manager::GatewayView` keeps resolving.

All 27 openclaw tests still pass.
