+++ apps/agent/src-tauri/src/openclaw/mod.rs
@@ -4,7 +4,22 @@ pub mod manager;
 pub mod protocol;
 pub mod secret_store;
 
-pub use manager::{
+mod backoff;
+mod command;
+mod connection;
+mod device_proof;
+mod dto;
+mod identity;
+mod runtime_state;
+mod session;
+mod url_validation;
+mod util;
+
+#[cfg(test)]
+mod tests;
+
+pub use dto::{
     GatewayDiscoverInput, GatewayListResponse, GatewayRequestInput, GatewayUpsertRequest,
-    ImportGatewayRequest, ImportGatewayResponse, OpenClawAgentEvent, OpenClawManager,
+    ImportGatewayRequest, ImportGatewayResponse, OpenClawAgentEvent,
 };
+pub use manager::OpenClawManager;
